backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.0.339
langchain-google-genai==0.0.6
langchain-community==0.0.10
kiteconnect==4.1.0
sqlalchemy==2.0.23
pandas==2.1.4
requests==2.31.0
pathlib==1.0.1
tavily-python==0.3.3 